+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add turn length vs turn number plot to `analysis/SampleAnalysis-SingleGame.ipynb`.
 * Reorder game initilization so that most of the info can be entered before the player order is determined.
 * Announce gameID and location at the start.
+* Status printout now shows time elapsed since end of previous turn.
 
 ### 0.4.1 (18 November 2018)